Add angle class and move some angle calculations there

+++ WarriorsSnuggery.Game/Position/Vector.cs
@@ -7,7 +7,7 @@ namespace WarriorsSnuggery
 	{
 		public static readonly Vector Zero = new Vector();
 
-		public float Dist => (float)Math.Sqrt(X * X + Y * Y + Z * Z);
+		public float Dist => (float)Math.Sqrt((double)X * X + (double)Y * Y + (double)Z * Z);
 
 		public readonly float X;
 		public readonly float Y;
